[integration] Show stderr/stdout from failed subprocesses
Hook into pytest so that when a CalledProcessError is thrown from either
building or running the CLI or integration-stub, we display the process's
stdout and stderr and suppress the noisy Python stack trace.

For errors when running Rust binaries, we monkey-patch pyunit's Item.runtest
for lack of a working hook to accomplish this.  This way we can still use
pytest.raises() to catch and inspect desired exceptions in the test cases as
usual, which would not be possible if we wrapped subprocess.run directly.

For errors from `cargo build`, we wrap its `subprocess.run` call in cargo.py.