[integration] Cache conversion, message reading
Some tests are of the form "do a simple conversion of this GPX file into
FIT, and then read the FIT messages in its output with the Garmin SDK.
Because conversion parsing take time, we can cache the read messages
between test cases where the input stays the same.

This seems to significantly speed up tests where it's applicable, at the
cost of hopefully not too much obfuscation of the commands under test.