[course] Allow specifying intercept threshold
Adds a Course option and a corresponding CLI argument for the interception
threshold, which is how close the course has to come to a waypoint for it to be
considered a "course point".

This change also changes course point computation so that points are written to
the course in order of their distance.