[course] Put GeoAndXyzPoint in Waypoint
This change relies on making a bunch of the algorithm and type-module
code generic on the point types.  This is accomplished by adding a
HasXyzPoint trait analogous to what's now named HasGeoPoint.  Some of
the generics are a bit messy with regard to dereferencing, and might
need to be cleaned up.