[course] Add "floor" feature
Adds a feature, enabled by default, to choose whether we use
intercept_distance_floor to filter out "Far" intercepts.  This works by
adding a type parameter on the GeoSegment and Waypoint point type to
CourseSetBuilder and friends.

At the moment this is pretty ugly because it pollutes the code with
extra type constraints, which I'll want to find a way to work around.
Also has a bunch of clippy warnings right now.