[course] Refactor CourseBuilder/SegmentedCourseBuilder
Splits the old CourseBuilderImpl into two types to allow for a two-stage
build.  The first stage (segmentation) does the work of solving the
indirect problem between adjacent points and geocentric (XYZ) forward if
necessary, while the second converts into a Course, after the
CourseSetBuilderImpl has had the chance to run process_waypoint and
append course points to the SegmentedCourseBuilder.