[course] Hold references to GeoPoint in GeoSegment
Changes GeoSegment to hold lifetimed references to the underlying
GeoPoints instead of creating redundant copies.  This removes
duplication of GeoAndXyzPoint in GeoSegment (now the segment holds
pointers instead), and it provides a slight 5-10ms speedup on the stress
test gpx006 course.

We're still copying a bunch of data when lifting GeoPoints into
GeoAndXyzPoints, but it would be a pain to plumb the lifetimes needed to
work around that through CourseSetBuilderImpl.