[devtools] Rename the integration-stub package to devtools
We keep integration-stub's binary with the same name.  This gives us a
place to put another binary for random command-line debugging stuff that
doesn't belong in the main CLI or package, and probably also a good
place to merge the Python tools/ contents (as opposed to scripts).

Just a little tidier...