[ffi] Write strings back to caller
Modifies the FFI so that when getting version strings for GeographicLib or the
C++ compiler, we now write them back to a caller-provided buffer.  This will
allow us to hook these up with a JavaScript shim if we use GeographicLib and
CoursePointer compiled to two different WASM modules that don't share memory.